revert ES6-style import in content/background scripts
This reverts commit 84680f310dc7b38bcafea18fde5ba83a537671ab.
