--- packages/lexical-devtools/src/background/index.html
@@ -1,1 +1,0 @@
-<script type="module" src="./index.ts"></script>
