+++ packages/lexical-devtools/src/background/index.ts
@@ -5,20 +5,18 @@
  * LICENSE file in the root directory of this source tree.
  *
  */
-import {FROM_APP, FROM_CONTENT} from '../modules';
-
 // Create messaging connection to send editorState updates to Lexical DevTools App.
 const ports = {}; // Each tab will have a separate messaging port for the devTools app & the inspectedWindow's content script, eg. { tabId: { reactPort, contentScriptPort } }
 
 // The Lexical DevTools React UI sends a message to initialize the port.
 chrome.runtime.onConnect.addListener(function (port) {
   port.onMessage.addListener((message) => {
-    if (message.name === 'init' && message.type === FROM_APP) {
+    if (message.name === 'init' && message.type === 'FROM_APP') {
       ports[message.tabId].react = port;
       return;
     }
 
-    if (message.name === 'init' && message.type === FROM_CONTENT) {
+    if (message.name === 'init' && message.type === 'FROM_CONTENT') {
       ports[port.sender.tab.id] = {};
       ports[port.sender.tab.id].content = port;
       return;
