--- packages/lexical-devtools/src/modules/index.ts
@@ -1,13 +1,0 @@
-/**
- * Copyright (c) Meta Platforms, Inc. and affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- *
- */
-
-const FROM_APP = 'FROM_APP';
-const FROM_CONTENT = 'FROM_CONTENT';
-const FROM_PAGE = 'FROM_PAGE';
-
-export {FROM_APP, FROM_CONTENT, FROM_PAGE};
